Make VoteScore an object

+++ src/main/java/com/github/kory33/signvote/model/VotePointStats.java
@@ -36,7 +36,7 @@ public class VotePointStats {
 
         this.voteSession.getVoteManager().getVotes(this.votePoint)
             .forEach(vote -> {
-                int score = vote.getScore();
+                int score = vote.getScore().toInt();
                 int count = distrMap.getOrDefault(score, 0);
                 distrMap.put(score, count + 1);
             });
